cambio de arquitectura con Capa de Dominio
