--- demo/src/main/java/com/pineapple/demo/models/controller/Principal.java
@@ -1,19 +1,0 @@
-
-package com.pineapple.demo.models.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.servlet.ModelAndView;
-
-
-@Controller
-public class Principal {
-
-    @GetMapping("/")
-    public ModelAndView inicio() {
-        return new ModelAndView("index");
-    }
-}
-    
-    
-
