--- demo/src/main/java/com/pineapple/demo/models/entity/Product.java
@@ -1,114 +1,0 @@
-package com.pineapple.demo.models.entity;
-
-import java.io.Serializable;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-
-@Entity
-@Table (name="products")
-public class Product implements Serializable {
-
-	@Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id_product")
-    private Long id;
-    private String productName;
-    private String description;
-    private Integer quantity;
-    @Column(name = "unit_price")
-    private Double unitPrice;
-    private String picture;
-    @ManyToOne
-    private ProductCategory productCategory;
-    
-    private static final long serialVersionUID = 1L;
-
-    public Product() {
-    }
-
-    public Product(String productName, String description, Integer quantity, Double unitPrice, String picture) {
-        this.productName = productName;
-        this.description = description;
-        this.quantity = quantity;
-        this.unitPrice = unitPrice;
-        this.picture = picture;
-        //this.productCategory = productCategory;
-    }
-
-    public String getProductName() {
-        return productName;
-    }
-
-    public void setProductName(String productName) {
-        this.productName = productName;
-    }
-
-    public ProductCategory getProductCategory() {
-        return productCategory;
-    }
-
-    public void setProductCategory(ProductCategory productCategory) {
-        this.productCategory = productCategory;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	public String getName() {
-		return productName;
-	}
-
-	public void setName(String productName) {
-		this.productName = productName;
-	}
-
-
-	public Integer getQuantity() {
-		return quantity;
-	}
-
-	public void setQuantity(Integer quantity) {
-		this.quantity = quantity;
-	}
-
-	public Double getUnitPrice() {
-		return unitPrice;
-	}
-
-	public void setUnitPrice(Double unitPrice) {
-		this.unitPrice = unitPrice;
-	}
-
-	public String getPicture() {
-		return picture;
-	}
-
-	public void setPicture(String picture) {
-		this.picture = picture;
-	}
-
-	public static long getSerialversionuid() {
-		return serialVersionUID;
-	}
-    
-}
