--- demo/src/main/java/com/pineapple/demo/models/security/AppConfig.java
@@ -1,15 +1,0 @@
-
-package com.pineapple.demo.models.security;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-
-@Configuration
-public class AppConfig {
-    
-    @Bean
-    public BCryptPasswordEncoder encoder(){
-    return new  BCryptPasswordEncoder();
-    }
-}
