+++ demo/src/main/java/com/pineapple/demo/persistence/entity/Product.java
@@ -1,0 +1,86 @@
+package com.pineapple.demo.persistence.entity;
+
+import java.io.Serializable;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+
+@Entity
+@Table (name="products")
+public class Product implements Serializable {
+
+	@Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id_product")
+    private Long idProduct;
+    @Column(name = "product_name")
+    private String productName;
+    private String description;
+    private Integer quantity;
+    @Column(name = "unit_price")
+    private Double unitPrice;
+    //private String picture;
+    @ManyToOne
+    @JoinColumn(name="product_category", insertable = false, updatable =false )
+    private Category productCategory;
+    
+    private static final long serialVersionUID = 1L;
+
+    //constructor
+
+
+    //getters and setters
+    public Long getIdProduct() {
+        return idProduct;
+    }
+
+    public void setIdProduct(Long idProduct) {
+        this.idProduct = idProduct;
+    }
+
+    public String getProductName() {
+        return productName;
+    }
+
+    public void setProductName(String productName) {
+        this.productName = productName;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public Integer getQuantity() {
+        return quantity;
+    }
+
+    public void setQuantity(Integer quantity) {
+        this.quantity = quantity;
+    }
+
+    public Double getUnitPrice() {
+        return unitPrice;
+    }
+
+    public void setUnitPrice(Double unitPrice) {
+        this.unitPrice = unitPrice;
+    }
+
+    public Category getProductCategory() {
+        return productCategory;
+    }
+
+    public void setProductCategory(Category productCategory) {
+        this.productCategory = productCategory;
+    }
+}
