+++ demo/src/main/java/com/pineapple/demo/persistence/entity/Users.java
@@ -1,4 +1,4 @@
-package com.pineapple.demo.models.entity;
+package com.pineapple.demo.persistence.entity;
 
 import java.io.Serializable;
 import java.util.Date;
