+++ demo/src/test/java/com/pineapple/demo/StartTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class DemoApplicationTests {
+class StartTests {
 
 	@Test
 	void contextLoads() {
